Added "unsigned long" version of saturate_cast.

+++ modules/core/include/opencv2/core/operations.hpp
@@ -118,7 +118,7 @@ template<typename _Tp> static inline _Tp saturate_cast(schar v) { return _Tp(v);
 template<typename _Tp> static inline _Tp saturate_cast(ushort v) { return _Tp(v); }
 template<typename _Tp> static inline _Tp saturate_cast(short v) { return _Tp(v); }
 template<typename _Tp> static inline _Tp saturate_cast(unsigned v) { return _Tp(v); }
-template<typename _Tp> static inline _Tp saturate_cast(size_t v) { return _Tp(v); }
+template<typename _Tp> static inline _Tp saturate_cast(unsigned long v) { return _Tp(v); }
 template<typename _Tp> static inline _Tp saturate_cast(int v) { return _Tp(v); }
 template<typename _Tp> static inline _Tp saturate_cast(float v) { return _Tp(v); }
 template<typename _Tp> static inline _Tp saturate_cast(double v) { return _Tp(v); }
